Add babble.bootstrap flag
The bootstrap flag tells babble to load its Store from a badger_db
file if it exists. This will replay all Babble blocks from the
existing database, thereby reseting the node to where it was when it
was last shutdown. If the flag is not set, Babble will create a
backup of the latest badger_db, and create a new one.