If we are intending that end users are sheilded from the underlying evm-lite and babble components we need to ensure that basic functions for them are available from within monetd. Take tests for example.
This commit adds make testevml to the makefile. This runs the test from
within the vendor folder copy of evm-lite.
make test is now a composite recipe of testmonetd, testevml and
testbabble. The script output is passed through sed to reduce the path
length as the vendor folder paths are so long that they render the
output fairly unreadable.

Babble tests are currently unavailable due to unfulfilled glide
dependencies, so the testbabble recipe is commented out.