Use same address for evm-lite and babble APIs
evm-lite and babble used to expose their respective APIs on
different ports. I removed the correponding configuration options,
and replaced it with a single api-address value. So they are both
serving on the same port now.

The monetd config pull command takes an address in the form
"ip:port". All the files are downloaded from that single URL base,
although the requests will be handled by both modules
independently.

This required a few changes on Babble and EVM-Lite so make sure to
pull the latest versions of the develop branches.

I also started to do some refactoring of the default configuration
values. Moved them from poa/common to configuration/.