Use block hash to pseudo-randomly pick coinbase
Replace round-robin mechanism with pseudo-random system using the
block hash. We derive a uint32 from the block hash, and compute the
modulo with respect to the size of the validator-set, to obtain the
index of the next validator who's address will be the coinbase.
This is a fair system because pseudo-random, and safe because it
should not be possible for malicious validators to manipulate the
block hash.