Don't log err when monetd config cannot find default key
This error message was confusing. The monetd CLI tries to provide
a default --key value for the config pull cmd. This default value is
computed upon initialisation of monetd, even if a user is actually
calling another command than config pull. The default --key value is
obtained by attempting to read the default keystore (~/.monet/keystore)
which is not necessarily readable by the current user. So, in the
eventuality that a user has no read access to the default keystore,
and overrides the default datadir option, an error message would
appear about trying to open ~/.monet/keystore without permissions.
We just removed the error message, it is ok for the default value
to be empty; the config pull function will actually use the first
key found in the specified datadir/keystore.