Add giverny parse to the giverny reference docs
Regenerate the screenshots used in the docs.

Also amend the generatedocsincluded to map $HOME to /home/user
This means that we do not expose our local user ids, and it means
that get the same result no matter who runs it.  As an example
most of the files in this commit, are solely due to a user name
change - albeit in this case to /home/user the new standard value.